Move everything into AppRouterView so we can use async permissions in the context sidebar.
