+++ orion-ui/src/App.vue
@@ -1,84 +1,14 @@
 <template>
   <div class="app" data-teleport-target="app">
-    <template v-if="!media.lg">
-      <PGlobalSidebar class="app__mobile-menu">
-        <template #upper-links>
-          <p-icon icon="PrefectGradient" class="app__prefect-icon" />
-        </template>
-        <template #bottom-links>
-          <PIcon icon="MenuIcon" class="app__menu-icon" @click="toggle" />
-        </template>
-      </PGlobalSidebar>
-    </template>
-    <ContextSidebar v-if="showMenu" class="app__sidebar" @click="close" />
     <suspense>
-      <AppRouterView class="app__router-view" />
+      <AppRouterView />
     </suspense>
   </div>
 </template>
 
 <script lang="ts" setup>
-  import { PGlobalSidebar, PIcon, media } from '@prefecthq/prefect-design'
-  import { computed, ref, watchEffect } from 'vue'
-  import ContextSidebar from '@/components/ContextSidebar.vue'
   import AppRouterView from '@/pages/AppRouterView.vue'
   import { healthCheck } from '@/utilities/api'
 
-  const mobileMenuOpen = ref(false)
-  const showMenu = computed(() => media.lg || mobileMenuOpen.value)
-
-  function toggle(): void {
-    mobileMenuOpen.value = !mobileMenuOpen.value
-  }
-
-  function close(): void {
-    mobileMenuOpen.value = false
-  }
-
   healthCheck()
-
-  watchEffect(() => document.body.classList.toggle('body-scrolling-disabled', showMenu.value && !media.lg))
 </script>
-
-<style>
-.body-scrolling-disabled { @apply
-  overflow-hidden
-}
-
-.app { @apply
-  text-slate-900
-  flex
-  flex-col
-}
-
-.app {
-  --prefect-scroll-margin: theme('spacing.20');
-  min-height: 100vh;
-}
-
-.app__prefect-icon { @apply
-  w-6
-  h-6
-}
-
-.app__menu-icon { @apply
-  text-white
-  w-6
-  h-6
-  cursor-pointer
-}
-
-@screen lg {
-  .app {
-    --prefect-scroll-margin: theme('spacing.2');
-    display: grid;
-    grid-template-columns: max-content minmax(0, 1fr);
-  }
-}
-
-.app__router-view {
-  /* The 1px flex-basis is important because it allows us to use height: 100% without additional flexing */
-  flex: 1 0 1px;
-  height: 100%;
-}
-</style>
